Ability to show updated board after a puzzle move / show hints
